+++ src/include/cd32_fmv.h
@@ -9,7 +9,7 @@ extern void cd32_fmv_state(int state);
 extern void cd32_fmv_new_image(int, int, int, uae_u8*);
 extern void cd32_fmv_genlock(struct vidbuffer*, struct vidbuffer*);
 extern void cd32_fmv_new_border_color(uae_u32);
-extern void cd32_fmv_set_sync(double svpos, double adjust);
+extern void cd32_fmv_set_sync(float svpos, float adjust);
 
 extern int cd32_fmv_active;
